Step 13.8: Implement newUser and findByUsername
